Fix symbol pruning in set operation nodes
Before this change, ExceptNode and IntersectNode were incorrectly
handled in PruneUnreferencedOutputs optimizer.
The symbols on which the set operations should be performed, were
incorrectly pruned in the case when they were not referenced
in the parent node. This could result in changed semantics of
the node.
Example:
- ProjectNode (x)
	- IntersectNode (x, y)
		- Source1 layout: (a, b)
		  values: (1, 2)
		- Source2 layout: (c, d)
		  values: (1, 3)
The IntersectNode should produce empty result.
However, if output symbol `y` was pruned, and source symbols `b` and `d`
were pruned accordingly, we get:
- ProjectNode (x)
	- IntersectNode (x)
		- Source1 layout: (a)
		  values: (1)
		- Source2 layout: (c)
		  values: (1)
Now the result of intersection is not empty.

Note: the incorrect code was never reached, because ExceptNode
and IntersectNode were rewritten on earlier stage of optimization.